Added Comments and Connected navigation from main page

+++ backend/reactions_app/views.py
@@ -5,6 +5,7 @@ from rest_framework.status import (
     HTTP_201_CREATED,
     HTTP_400_BAD_REQUEST,
     HTTP_204_NO_CONTENT,
+    HTTP_401_UNAUTHORIZED
 )
 from manateewatch_proj.settings import env
 from user_app.views import TokenReq
@@ -12,7 +13,7 @@ from map_app.models import Sighting_Data
 from django.shortcuts import get_object_or_404
 from profile_app.models import User_Profile
 from .models import Reactions
-from .serializers import CommentSerializer
+from .serializers import CommentSerializer, NewCommentSerializer
 
 
 # Create your views here.
@@ -22,23 +23,28 @@ class AllComments(TokenReq):
         data = request.data.copy()
         data["sighting"] = sighting_id
         data['user'] = request.user.id
-        ser_data = CommentSerializer(data=data)
+        ser_data = NewCommentSerializer(data=data)
         if ser_data.is_valid():
-            ser_data.save()
-            return Response(data, status=HTTP_201_CREATED)
+            new_comment = ser_data.save()
+            newPost = CommentSerializer(new_comment).data
+            return Response(newPost, status=HTTP_201_CREATED)
         print(ser_data.errors)
         return Response(ser_data.errors, status=HTTP_400_BAD_REQUEST)
     
 class A_Comment(TokenReq):
     '''allow comment author to delete their own comments and original sighting author to delete any comments on their post'''
-    def delete(self, request, id):
-        comment = get_object_or_404(Reactions, id=id)
-        comment.delete()
-        return Response('comment was deleted', status=HTTP_204_NO_CONTENT)
-    
+    def delete(self, request, comment_id, sighting_id):
+        user = request.user.id
+        comment = get_object_or_404(Reactions, id=comment_id)
+        commentData = NewCommentSerializer(comment).data
+        print(commentData['user'], user)
+        if (user == commentData['user']):
+            comment.delete()
+            return Response('comment was deleted', status=HTTP_204_NO_CONTENT)
+        return Response("Unauthorized access", status=HTTP_401_UNAUTHORIZED)
     '''allow users to edit their own comments'''
-    def put(self, request, id):
-        comment = get_object_or_404(Reactions, id=id)
+    def put(self, request, comment_id, sighting_id):
+        comment = get_object_or_404(Reactions, id=comment_id)
         data = request.data.copy()
         edit_comment = CommentSerializer(comment, data=data, partial=True)
         if edit_comment.is_valid():
